Some re-arrangement of the functions
Computing the individual voronoi polygons is now a `method_util` and not part of the density calculator
