+++ pedpy/methods/method_utils.py
@@ -1,10 +1,12 @@
 """Helper functions for the analysis methods."""
 import itertools
-from typing import Tuple
+from collections import defaultdict
+from typing import List, Optional, Tuple
 
 import numpy as np
 import pandas as pd
 import shapely
+from scipy.spatial import Voronoi
 from shapely import LineString, Polygon
 
 from pedpy.data.geometry import Geometry
@@ -184,6 +186,166 @@ def compute_neighbors(individual_voronoi_data: pd.DataFrame) -> pd.DataFrame:
         neighbor_df.append(frame_df)
 
     return pd.concat(neighbor_df)
+
+
+def compute_individual_voronoi_polygons(
+    *,
+    traj_data: pd.DataFrame,
+    geometry: Geometry,
+    cut_off: Optional[Tuple[float, int]] = None,
+    use_blind_points: bool = False,
+) -> pd.DataFrame:
+    """Compute the individual voronoi cells for each person and frame.
+
+    Args:
+        traj_data (pd.DataFrame): trajectory data
+        geometry (Geometry): bounding area, where pedestrian are supposed to be
+        cut_off (Tuple[float, int]): radius of max extended voronoi cell (in
+                m), number of linear segments in the approximation of circular
+                arcs, needs to be divisible by 4!
+        use_blind_points (bool): adds extra 4 points outside the geometry to
+                also compute voronoi cells when less than 4 peds are in the
+                geometry
+
+    Returns:
+        DataFrame containing the columns: 'ID', 'frame' and 'individual
+        voronoi'.
+    """
+    dfs = []
+
+    bounds = geometry.walkable_area.bounds
+    x_diff = abs(bounds[2] - bounds[0])
+    y_diff = abs(bounds[3] - bounds[1])
+    clipping_diameter = 2 * max(x_diff, y_diff)
+
+    if use_blind_points:
+        blind_points = np.array(
+            [
+                [100 * (bounds[0] - x_diff), 100 * (bounds[1] - y_diff)],
+                [100 * (bounds[2] + x_diff), 100 * (bounds[1] - y_diff)],
+                [100 * (bounds[0] - x_diff), 100 * (bounds[3] + y_diff)],
+                [100 * (bounds[2] + x_diff), 100 * (bounds[3] + y_diff)],
+            ]
+        )
+
+    for _, peds_in_frame in traj_data.groupby(traj_data.frame):
+        points = peds_in_frame[["X", "Y"]].values
+        if use_blind_points:
+            points = np.concatenate([points, blind_points])
+
+        # if blind points were added, len(points) will always be > 4
+        if len(points) < 4:
+            continue
+
+        vor = Voronoi(points)
+        voronoi_polygons = _clip_voronoi_polygons(vor, clipping_diameter)
+
+        if use_blind_points:
+            voronoi_polygons = voronoi_polygons[:-4]
+        voronoi_in_frame = peds_in_frame.loc[:, ("ID", "frame", "points")]
+
+        # Compute the intersecting area with the walkable area
+        voronoi_in_frame["individual voronoi"] = shapely.intersection(
+            voronoi_polygons, geometry.walkable_area
+        )
+
+        if cut_off is not None:
+            num_edges = cut_off[1]
+            radius = cut_off[0]
+            quad_edges = int(num_edges / 4)
+            voronoi_in_frame["individual voronoi"] = shapely.intersection(
+                voronoi_in_frame["individual voronoi"],
+                shapely.buffer(
+                    peds_in_frame["points"], radius, quad_segs=quad_edges
+                ),
+            )
+
+        # Only consider the parts of a multipolygon which contain the position
+        # of the pedestrian
+        voronoi_in_frame.loc[
+            shapely.get_type_id(voronoi_in_frame["individual voronoi"]) != 3,
+            "individual voronoi",
+        ] = voronoi_in_frame.loc[
+            shapely.get_type_id(voronoi_in_frame["individual voronoi"]) != 3, :
+        ].apply(
+            lambda x: shapely.get_parts(x["individual voronoi"])[
+                shapely.within(
+                    x["points"], shapely.get_parts(x["individual voronoi"])
+                )
+            ][0],
+            axis=1,
+        )
+
+        dfs.append(voronoi_in_frame)
+
+    return pd.concat(dfs)[["ID", "frame", "individual voronoi"]]
+
+
+def _clip_voronoi_polygons(
+    voronoi: Voronoi, diameter: float
+) -> List[shapely.Polygon]:
+    """Generate Polygons from the Voronoi diagram.
+
+    Generate shapely.geometry.Polygon objects corresponding to the
+    regions of a scipy.spatial.Voronoi object, in the order of the
+    input points. The polygons for the infinite regions are large
+    enough that all points within a distance 'diameter' of a Voronoi
+    vertex are contained in one of the infinite polygons.
+    from: https://stackoverflow.com/a/52727406/9601068
+    """
+    polygons = []
+    centroid = voronoi.points.mean(axis=0)
+
+    # Mapping from (input point index, Voronoi point index) to list of
+    # unit vectors in the directions of the infinite ridges starting
+    # at the Voronoi point and neighbouring the input point.
+    ridge_direction = defaultdict(list)
+    for (p, q), rv in zip(voronoi.ridge_points, voronoi.ridge_vertices):
+        u, v = sorted(rv)
+        if u == -1:
+            # Infinite ridge starting at ridge point with index v,
+            # equidistant from input points with indexes p and q.
+            t = voronoi.points[q] - voronoi.points[p]  # tangent
+            n = np.array([-t[1], t[0]]) / np.linalg.norm(t)  # normal
+            midpoint = voronoi.points[[p, q]].mean(axis=0)
+            direction = np.sign(np.dot(midpoint - centroid, n)) * n
+            ridge_direction[p, v].append(direction)
+            ridge_direction[q, v].append(direction)
+
+    for i, r in enumerate(voronoi.point_region):
+        region = voronoi.regions[r]
+        if -1 not in region:
+            # Finite region.
+            polygons.append(shapely.polygons(voronoi.vertices[region]))
+            continue
+        # Infinite region.
+        inf = region.index(-1)  # Index of vertex at infinity.
+        j = region[(inf - 1) % len(region)]  # Index of previous vertex.
+        k = region[(inf + 1) % len(region)]  # Index of next vertex.
+        if j == k:
+            # Region has one Voronoi vertex with two ridges.
+            dir_j, dir_k = ridge_direction[i, j]
+        else:
+            # Region has two Voronoi vertices, each with one ridge.
+            (dir_j,) = ridge_direction[i, j]
+            (dir_k,) = ridge_direction[i, k]
+
+        # Length of ridges needed for the extra edge to lie at least
+        # 'diameter' away from all Voronoi vertices.
+        length = 2 * diameter / np.linalg.norm(dir_j + dir_k)
+
+        # Polygon consists of finite part plus an extra edge.
+        finite_part = voronoi.vertices[region[inf + 1 :] + region[:inf]]
+        extra_edge = np.array(
+            [
+                voronoi.vertices[j] + dir_j * length,
+                voronoi.vertices[k] + dir_k * length,
+            ]
+        )
+        polygons.append(
+            shapely.polygons(np.concatenate((finite_part, extra_edge)))
+        )
+    return polygons
 
 
 def _compute_individual_movement(
